move walk in utils and extractID in test-cases
